Update blocklayered.js
i was having problems with dropdown because event was triggered by click. it should be triggered by "change"
i can see thats been changed but i havent tried this new version, and i dont know if the issue was fixed, so...
NOTE! I havent tested this yet! I dont even know if it was already fixed in the dev code, so please check it your self

i dont know if it makes a difference using the "body" selector but it might

found most of the code here:
http://forge.prestashop.com/browse/PNM-2444

sorry this is a quickie :)