Update blocklayered_admin.js
A known bug of blocklayered is that it doesn't hold a changed order of the filters. You can drag and drop the lines in the backoffice and it will work. However, when you re-open the filter-template it will show you the elements in the original order. Saving that order will undo your sorting. See for example https://www.prestashop.com/forums/topic/317265-layered-navigation-block-dont-keep-order-position/ and https://www.prestashop.com/forums/topic/336943-layered-navigation-block-filters-order/). This modification repairs that problem. 

As Prestashop saves only positions of the active filters in this template we can only reconstruct that. I have implemented it at the end of the filter list.